Refactor code
- remove unnecessary code
- moving package
