+++ src/main/java/com/yam/app/account/infrastructure/LoginAccount.java
@@ -1,13 +1,10 @@
-package com.yam.app.account.presentation;
+package com.yam.app.account.infrastructure;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-/**
- * 현재 로그인 되어있는 사용자의 세션 AccountPrincipal 객체를 반환합니다.
- */
 @Target(ElementType.PARAMETER)
 @Retention(RetentionPolicy.RUNTIME)
 public @interface LoginAccount {
