+++ src/main/java/com/yam/app/account/presentation/AccountQueryApi.java
@@ -2,8 +2,8 @@ package com.yam.app.account.presentation;
 
 import com.yam.app.account.application.AccountFacade;
 import com.yam.app.account.infrastructure.AccountPrincipal;
+import com.yam.app.account.infrastructure.LoginAccount;
 import javax.validation.Valid;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
@@ -13,7 +13,6 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-@Slf4j
 @RestController
 @RequestMapping(
     produces = MediaType.APPLICATION_JSON_VALUE,
